+++ src/pages/profile/profile_edit/profile_edit.js
@@ -1,6 +1,6 @@
 import '@/styles/pages/profile.scss';
 import '@/layout/footer';
-import { renderProfileItem } from '@/layout/profile_item';
+import { renderProfileItem } from '@/pages/profile/profile_item';
 import { getNode } from '@/library/index';
 
 renderProfileItem('edit');
